+++ Backend/package.json
@@ -16,5 +16,8 @@
     "express": "^4.18.2",
     "mysql2": "^3.3.1",
     "nodemon": "^2.0.22"
+  },
+  "devDependencies": {
+    "eslint": "^8.41.0"
   }
 }
